LPS-160648 Avoid code repetition when using useContext in Panel component
