+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/Panel/Panel.tsx
@@ -15,10 +15,13 @@
 import {ClayButtonWithIcon} from '@clayui/button';
 import ClayIcon from '@clayui/icon';
 import classNames from 'classnames';
-import React, {useContext} from 'react';
+import React from 'react';
 
 import {BoxType} from '../Layout/types';
-import PanelContextProvider, {PanelContext, TYPES} from './context';
+import PanelContextProvider, {
+	TYPES,
+	usePanelContext,
+} from './objectPanelContext';
 
 import './Panel.scss';
 
@@ -42,7 +45,7 @@ const Panel: React.FC<React.HTMLAttributes<HTMLElement>> & {
 interface IPanelBodyProps extends React.HTMLAttributes<HTMLElement> {}
 
 const PanelBody: React.FC<IPanelBodyProps> = ({children, className}) => {
-	const [{expanded}] = useContext(PanelContext);
+	const [{expanded}] = usePanelContext();
 
 	return (
 		<>
@@ -73,7 +76,7 @@ const PanelHeader: React.FC<IPanelHeaderProps> = ({
 	title,
 	type,
 }) => {
-	const [{expanded}, dispatch] = useContext(PanelContext);
+	const [{expanded}, dispatch] = usePanelContext();
 
 	return (
 		<div
